Implement object renaming
In this commit the name buffer is always reallocated, even if it's big enough for the new name. Fix in next commit.

OSCAudioBase.cpp - implement renameObject(), and make it available even in an otherwise static audio environment
OSCAudioBase.h - refactor name setting out to its own function, so we can use it both at construction time and when renaming; move stuff about so the "dynamic" interface is available for renaming
OSCAudioSend.py - rename the mixer object and send it stuff, then put it back