Improve immunity to "missing" audio objects
Some audio objects aren't implemented in some Teensy models, which tends to be "controlled" by some conditional compilation in header and/or cpp files. This is an attempt to detect the missing objects by spotting the absence of the #include guard defines. The relevant OSCAudio objects are likewise omitted, and an even more kludgy hack used to exclude the objects from the OSC_AUDIO_CLASSES macro!