First try at run-time resource clash checks
Normally done by the GUI, but can be bad if other systems bypass the checks, leading to crashes etc. This version assumes that once created, even if it's destroyed there will be some remnant of the I/O object consuming the resource. If we had a proper destructor then this could be dispensed with, but would also need a system of links or resource counters to be sure when a resource's setting could be reset to "unused".